backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "bs4>=0.0.2",
    "docx>=0.2.4",
    "fastapi>=0.121.0",
    "google-genai>=1.47.0",
    "lxml[html-clean]>=6.0.1",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.3",
    "pymongo>=4.6.0",
    "pymupdf>=1.26.6",
    "pymupdf4llm>=0.1.9",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "pandas>=2.2.2",
    "yfinance>=0.2.44",
    "jose>=1.0.0",
    "motor>=3.7.1",
]

[dependency-groups]
dev = [
    "uvicorn>=0.38.0",
]
